✅ Оновлення: вся структура перенесена в папку lec02, оновлений README
