--- job_1/test_request.py
@@ -1,44 +1,0 @@
-import os
-import requests
-
-
-def main() -> None:
-    """
-    Надсилає POST-запит до локального Flask-сервера (порт 8081),
-    щоб запустити джобу завантаження даних з API.
-    """
-
-    # Отримуємо абсолютний шлях до папки, де будуть зберігатися сирі JSON-файли.
-    # Абсолютний шлях — це повний шлях від кореня файлової системи.
-    # Наприклад:
-    #   - Відносний шлях: 'raw/sales'
-    #   - Абсолютний шлях (Windows): 'C:/Users/Andrii/Documents/GitHub/Data-Engeneer-tasks/raw/sales'
-    # Абсолютні шляхи гарантують, що програма точно знайде потрібну папку, незалежно від того,
-    # з якої директорії запускається код.
-    raw_dir: str = os.path.abspath("raw/sales")
-
-    # Формуємо JSON-об'єкт з параметрами для запиту.
-    # У нашому випадку ми передаємо тільки один параметр — шлях до директорії збереження.
-    data: dict[str, str] = {
-        "raw_dir": raw_dir
-    }
-
-    # Вказуємо адресу локального Flask-сервера.
-    # Порт 8081 використовується у першій джобі (fetch_sales).
-    url: str = "http://127.0.0.1:8081/run-job"
-
-    # Надсилаємо POST-запит до Flask-сервера з тілом запиту у форматі JSON.
-    response = requests.post(url, json=data)
-
-    # Виводимо HTTP-статус відповіді (наприклад, 200 — успіх, 500 — помилка сервера).
-    print("Відповідь від сервера:")
-    print(response.status_code)
-
-    # Виводимо тіло відповіді сервера (повідомлення про успіх або помилку).
-    print(response.json())
-
-
-# Стандартна конструкція для запуску скрипту напряму (а не при імпорті в інші модулі)
-if __name__ == "__main__":
-    main()
-
