+++ lec02/job_1/fetch_sales.py
