+++ lec02/job_1/test_request.py
@@ -1,0 +1,44 @@
+import os
+import requests
+
+
+def main() -> None:
+    """
+    Надсилає POST-запит до локального Flask-сервера (порт 8081),
+    щоб запустити джобу завантаження даних з API.
+    """
+
+    # Отримуємо абсолютний шлях до папки, де будуть зберігатися сирі JSON-файли.
+    # Абсолютний шлях — це повний шлях, який однозначно вказує на розташування
+    # директорії в файловій системі, незалежно від поточної робочої директорії.
+    # Наприклад:
+    #   - Відносний шлях: 'raw/sales'
+    #   - Абсолютний шлях (Windows): 'C:/Users/Andrii/Documents/GitHub/Data-Engeneer-tasks/raw/sales'
+    raw_dir: str = os.path.abspath("lec02/raw/sales")
+
+    # Формуємо тіло запиту у форматі JSON.
+    # Передаємо лише один параметр — абсолютний шлях до директорії для збереження.
+    data: dict[str, str] = {
+        "raw_dir": raw_dir
+    }
+
+    # Вказуємо адресу та порт, на якому працює локальний Flask-сервер для job_1.
+    # Цей маршрут "/run-job" відповідає за запуск джоби завантаження JSON-файлів.
+    url: str = "http://127.0.0.1:8081/run-job"
+
+    # Надсилаємо POST-запит до Flask-сервера, передаючи JSON-параметри.
+    response = requests.post(url, json=data)
+
+    # Виводимо HTTP-статус відповіді (наприклад: 200 — успіх, 400 — помилка клієнта, 500 — помилка сервера).
+    print("Відповідь від сервера:")
+    print(response.status_code)
+
+    # Виводимо JSON-відповідь сервера, яка містить повідомлення про успіх або помилку.
+    print(response.json())
+
+
+# Захищаємо запуск скрипту: цей блок виконується тільки при запуску напряму,
+# ігнорується при імпорті файлу як модуля в інші частини проєкту.
+if __name__ == "__main__":
+    main()
+
