+++ lec02/raw/sales/2022-08-09/sales_2022-08-09.json
