+++ lec02/raw/sales/2022-08-10/sales_2022-08-10.json
